separate common functions from normal source
