+++ content/download_section.py
@@ -3,7 +3,7 @@ import streamlit as st
 from pathlib import Path
 import shutil
 
-from src.common import page_setup
+from src.common.common import page_setup
 from zipfile import ZipFile, ZIP_DEFLATED
 
 page_setup()
