+++ content/run_subprocess.py
@@ -4,7 +4,7 @@ import os
 
 from pathlib import Path
 
-from src.common import page_setup, save_params
+from src.common.common import page_setup, save_params
 from src.run_subprocess import run_subprocess
 
 # Page name "workflow" will show mzML file selector in sidebar
