+++ content/simple_workflow.py
@@ -1,6 +1,6 @@
 import streamlit as st
 
-from src.common import page_setup, save_params, show_table
+from src.common.common import page_setup, save_params, show_table
 from src import simpleworkflow
 
 # Page name "workflow" will show mzML file selector in sidebar
