+++ content/topp_workflow_file_upload.py
@@ -1,5 +1,5 @@
 import streamlit as st
-from src.common import page_setup
+from src.common.common import page_setup
 from src.Workflow import Workflow
