+++ src/view.py
@@ -6,7 +6,7 @@ import plotly.graph_objects as go
 import streamlit as st
 import pyopenms as poms
 from src.plotting.MSExperimentPlotter import plotMSExperiment
-from src.common import show_fig
+from src.common.common import show_fig
 
 from typing import Union
